Move lib/labels into lib/async-requests
