+++ client/lib/async-requests/index.js
@@ -11,6 +11,11 @@ import { identity } from 'lodash';
 import { getIdsFromQuery, stringifyQuery } from '@woocommerce/navigation';
 
 /**
+ * Internal dependencies
+ */
+import { NAMESPACE } from 'store/constants';
+
+/**
  * Get a function that accepts ids as they are found in url parameter and
  * returns a promise with an optional method applied to results
  *
@@ -32,3 +37,40 @@ export function getRequestByIdString( path, handleData = identity ) {
 		return apiFetch( { path: pathString + payload } ).then( data => data.map( handleData ) );
 	};
 }
+
+export const getCategoryLabels = getRequestByIdString(
+	NAMESPACE + 'products/categories',
+	category => ( {
+		id: category.id,
+		label: category.name,
+	} )
+);
+
+export const getCouponLabels = getRequestByIdString( NAMESPACE + 'coupons', coupon => ( {
+	id: coupon.id,
+	label: coupon.code,
+} ) );
+
+export const getCustomerLabel = getRequestByIdString( NAMESPACE + 'customers', customer => ( {
+	id: customer.id,
+	label: customer.username,
+} ) );
+
+export const getProductLabels = getRequestByIdString( NAMESPACE + 'products', product => ( {
+	id: product.id,
+	label: product.name,
+} ) );
+
+export const getVariationLabels = getRequestByIdString(
+	query => NAMESPACE + `products/${ query.products }/variations`,
+	variation => {
+		return {
+			id: variation.id,
+			label: variation.attributes.reduce(
+				( desc, attribute, index, arr ) =>
+					desc + `${ attribute.option }${ arr.length === index + 1 ? '' : ', ' }`,
+				''
+			),
+		};
+	}
+);
